feat(tests): test caching
