+++ tests/integration/test_url.py
@@ -1,12 +1,6 @@
 import pytest
 
-
-async def _client_get(app, ufs):
-    client = app.test_client()
-    response = await client.get(ufs, headers=[("Accept-Encoding", "gzip")])
-    assert response.status_code == 200
-
-    return response
+from integration.helper import _client_get
 
 
 @pytest.mark.filterwarnings("ignore::RuntimeWarning")
